IMP extend testing utilities
Add an assertEqual variation that can ignore the order items
appear in mutable nested structures. Useful to compare DataFrames
that contain arrays of structs where the array was constructed
by grouping by, so order is irrelevant.

Add optional implicit test target importing in order to make
sure that SparkSession is not initialized as a side effect
of pyspark.sql.functions.udf before the test class can
initialize it