FIX support multiple sparkly sessions in tests
Allow a developer to employ a variety of sparkly sessions during testing, each
of which might have possibly different package requirements. To achieve this:

1. explicitly shut down the gateway and the underlying JVM before starting
a new session.

2. make sure SparklySession cleans up the environment so that the next session
has access to the same exact env and not to the modifed PYSPARK_SUBMIT_ARGS.

Before, we were just stopping the spark session / context, which doesn't
have any effect on the gateway / JVM, and as a result, new jars were not
being installed.