IMP remove has_package and has_jar functions
These functions don't actually inspect the environment, but rather depend on `packages` or `jars` being defined in the class. This is problematic for environments where bootstrapping is done via `spark-submit`, and the `assert` calls are merely syntactic sugar anyways. Just let the infrastructure throw the appropriate class-not-found exception if the cluster is not prepared properly.